requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.8
Django==5.1.4
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-standardized-errors==0.14.1
gunicorn==23.0.0
Markdown==3.7
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-decouple==3.8
sqlparse==0.5.3
